Corrected error
Previous code over-counted the number of messages in the co-occurrence network.  With the error I had a count of 150. This resulted in also over-counting the number of channels the messages were published in. The error is resolved in this code; the co-occurrence network is formed from 52 messages sent in 13 channels.